# server/api/swagger/protos/v1/todo.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "protos/v1/todo.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "TodoService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/todos": {
      "get": {
        "summary": "List todos with pagination",
        "operationId": "TodoService_ListTodos",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListTodosResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "TodoService"
        ]
      },
      "post": {
        "summary": "Create a new todo",
        "operationId": "TodoService_CreateTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateTodoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateTodoRequest"
            }
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    },
    "/v1/todos/{id}": {
      "get": {
        "summary": "Get a specific todo",
        "operationId": "TodoService_GetTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetTodoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TodoService"
        ]
      },
      "delete": {
        "summary": "Delete a todo",
        "operationId": "TodoService_DeleteTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteTodoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "TodoService"
        ]
      },
      "patch": {
        "summary": "Update a todo",
        "operationId": "TodoService_UpdateTodo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateTodoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TodoServiceUpdateTodoBody"
            }
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    }
  },
  "definitions": {
    "TodoServiceUpdateTodoBody": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "completed": {
          "type": "boolean"
        }
      },
      "title": "Update request"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateTodoRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "title": "Create request"
    },
    "v1CreateTodoResponse": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/v1Todo"
        }
      },
      "title": "Create response"
    },
    "v1DeleteTodoResponse": {
      "type": "object"
    },
    "v1GetTodoResponse": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/v1Todo"
        }
      }
    },
    "v1ListTodosResponse": {
      "type": "object",
      "properties": {
        "todos": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Todo"
          }
        },
        "nextPageToken": {
          "type": "string"
        }
      },
      "title": "List response"
    },
    "v1Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "completed": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "Todo represents a single todo item"
    },
    "v1UpdateTodoResponse": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/v1Todo"
        }
      },
      "title": "Update response"
    }
  }
}